requirements.txt
gspread==6.1.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-genai>=1.0.0
openai==1.30.1
groq==0.9.0
python-dotenv==1.0.1
